requirements.txt
ag2[openai]
python-dotenv
psycopg[binary]
loguru